net: shell: Add iface command
Move "iface" command to a separate file.
